// autogpt_platform/frontend/src/app/api/__generated__/models/graphExecution.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * AutoGPT Agent Server
 * This server is used to execute agents that are created by the AutoGPT system.
 * OpenAPI spec version: 0.1
 */
import type { GraphExecutionPresetId } from "./graphExecutionPresetId";
import type { AgentExecutionStatus } from "./agentExecutionStatus";
import type { GraphExecutionStats } from "./graphExecutionStats";
import type { GraphExecutionInputs } from "./graphExecutionInputs";
import type { GraphExecutionOutputs } from "./graphExecutionOutputs";

export interface GraphExecution {
  id?: string;
  user_id: string;
  graph_id: string;
  graph_version: number;
  preset_id?: GraphExecutionPresetId;
  status: AgentExecutionStatus;
  started_at: string;
  ended_at: string;
  stats: GraphExecutionStats;
  inputs: GraphExecutionInputs;
  outputs: GraphExecutionOutputs;
}